Read use_external_data_format from ORTConfig file (#1917)
When quantizing the models >2Gb, it's important to
set the flag use_external_data_format to 'true',
since otherwise the quantization will fail due to

`ValueError: Message onnx.ModelProto exceeds maximum protobuf size of 2GB`

However, currently there is no way to set the parameter
when using optimum-cli because there is no such
command option. Theoretically, it could be set when
using ORTConfig file with -c comman flag, because one
of the configuration parameters in it is use_external_data_format.
In fact, the optimum code ignores it and does not pass it
in quantize() function.

The goal of this change is to close this gap.